mnist-emp-project/requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
eval_type_backport==0.2.2
filelock==3.13.1
fonttools==4.59.1
fsspec==2024.6.1
importlib_resources==6.5.2
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scipy==1.13.1
simsimd==6.5.1
six==1.17.0
stringzilla==3.12.6
sympy==1.13.3
torch==2.8.0+cu126
torchvision==0.23.0+cu126
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.12.2
zipp==3.23.0
